Datahandler kan lave en masse Datauasge og random channel

+++ src/Handlers/BbrHandler.java
@@ -8,7 +8,7 @@ import java.util.ArrayList;
 public class BbrHandler {
     private static ArrayList<Bbr> allBbr = new ArrayList<>();
 
-        public static Bbr createBbrManuelt(int propertyNumber, String adress, int houseNumber, int zipCode, double heatedSquareMeters, String propertyType){
+        public static Bbr createBbrManuelt(int propertyNumber, String adress, int houseNumber, int zipCode, int heatedSquareMeters, String propertyType){
         Bbr createBbr = new Bbr(propertyNumber, adress, houseNumber,zipCode,heatedSquareMeters,propertyType);
         allBbr.add(createBbr);
         return createBbr;
